requirement.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
mysql-connector-python==8.2.0
python-engineio==4.8.1
python-socketio==5.10.0
simple-websocket==1.0.0
websockets==12.0
wsproto==1.2.0
